requirements.txt
language-tool-python>=2.7.0
pyspellchecker>=0.6.2
click>=8.1.0
